Release 0.7.1.1.0 (#293)
* improve rosenblatt docs

* safeguard rosenblatt

* fix @docType tag in pkg docs

* preprocessor define to fix BH issue (#270)


* allow bicop_dist() with tll (#268)

* improve docs for family = bicop_dist arguments

* allow bicop_dist() with tll

* export as.bicop()

---------


* fix na handling in vine() control checks (#266)


* Add website to DESCRIPTION (#271)

* Add website to DESCRIPTION

* `usethis::use_github_action("check-standard")`

* Create test-coverage.yaml

* Update R-CMD-check.yaml

* Update test-coverage.yaml

* fix handling of logistic distribution (#275)


* Add class of extreme value copulas and the tawn copula (#277)

* Added class of extreme value copulas and the tawn copula

* Changes to mainly parameters_to_tau

* add variable names to vinecop summary (#276)


* fix typo in docs (#281)


* Prep new version (#280)

* adapt to new cpp backend

* adapt discrete rosenblatt

* bump version

* pull vcl sources

* update website

* requirer higher kde1d version

* pull vcl

* final pull

* require new wdm

* update NEWS

* make CRAN happy

* more happiness

* final doc fix

* update API website

---------


* run reverse dependency checks on PRs to main (#288)

* run reverse dependency check on PRs to main

* trigger

* fix typo

---------


* fix typo in docs (fixes #273) (#289)


* add 'main' argument to pairs_copula_data() (fixes #278) (#290)

* add 'main' argument to pairs_copula_data() (fixes #278)

* document argument

---------


* enable fitting only parameters in vinecop() (#291)

* enable fitting only parameters in vinecop()

* warn about unused structure/family_set arguments

---------


* Pull latest vinecopulib (#292)

* update vinecoplib

* remove boost flags

* bump version

* pull vcl again

---------


* Better revdep check (#295)

* use base R revdep check via tools::

* another try

* separate check from results, disable mac

* try no-manual

* Don't check as CRAN

---------


* add allow_rotations argument (#296)

* pull cpp sources

* update R interface

* fix stuff

* shorten example runtime

---------


* updating website

* Improve coverage (#297)

* test variable name legend in summary(fit)

* remove now unused check_margin_controls() function

* test as.bicop

* test and improve bicop_dist("tll", ...)

* exclude vinecopulib sources from coverage computation

* remove unused interfacing functions

* test rvine_structure_sim()

* test structure validation

* update RcppExports

* not uploading?

* fix status badges

* htttps

---------


* Fix coverage upload (#298)

* fix codecov upload

* clean up on-push

* fix exclusions

* clean on-push  branches

---------


* Improve coverage further (#299)

* fix + improve unit tests for vine_dist()

* fix uncovered issues in vine_dist() logic

* test factor handling in vine()

* remove unused collate_u() function

* improve as_rvine_X coverage

* doc typo

---------


* don't use designated initializers (C++20 feature) (#300)


---------

+++ inst/include/vinecopulib/misc/tools_optimization.hpp
@@ -1,4 +1,4 @@
-// Copyright © 2016-2023 Thomas Nagler and Thibault Vatter
+// Copyright © 2016-2025 Thomas Nagler and Thibault Vatter
 //
 // This file is part of the vinecopulib library and licensed under the terms of
 // the MIT license. For a copy, see the LICENSE file in the root directory of
@@ -20,24 +20,24 @@ public:
 
   BobyqaControls(double initial_trust_region,
                  double final_trust_region,
-                 int maxeval);
+                 size_t maxeval);
 
   double get_initial_trust_region();
 
   double get_final_trust_region();
 
-  int get_maxeval();
+  size_t get_maxeval();
 
 private:
   double initial_trust_region_; //! Initial trust region
   double final_trust_region_;   //! Final trust region
-  int maxeval_; //! Maximal number of evaluations of the objective
+  size_t maxeval_; //! Maximal number of evaluations of the objective
 
   //! Sanity checks
   //! @{
   void check_parameters(double initial_trust_region,
                         double final_trust_region,
-                        int maxeval);
+                        size_t maxeval);
   //! @}
 };
 
@@ -49,7 +49,7 @@ public:
 
   void set_controls(double initial_trust_region,
                     double final_trust_region,
-                    int maxeval);
+                    size_t maxeval);
 
   Eigen::VectorXd optimize(
     const Eigen::VectorXd& initial_parameters,
